explicit casts

+++ crypto_kem/saber/clean/pack_unpack.c
@@ -8,7 +8,7 @@ void PQCLEAN_SABER_CLEAN_POLT2BS(uint8_t bytes[SABER_SCALEBYTES_KEM], const poly
     const uint16_t *in = data->coeffs;
     uint8_t *out = bytes;
     for (j = 0; j < SABER_N / 2; j++) {
-        out[0] = (in[0] & 0x0f) | (in[1] << 4);
+        out[0] = (uint8_t) ((in[0] & 0x0f) | (in[1] << 4));
         in += 2;
         out += 1;
     }
@@ -32,19 +32,19 @@ static void POLq2BS(uint8_t bytes[SABER_POLYBYTES], const poly *data) {
     const uint16_t *in = data->coeffs;
     uint8_t *out = bytes;
     for (j = 0; j < SABER_N / 8; j++) {
-        out[0] = in[0];
-        out[1] = ((in[0] >> 8) & 0x1f) | (in[1] << 5);
-        out[2] = in[1] >> 3;
-        out[3] = ((in[1] >> 11) & 0x03) | (in[2] << 2);
-        out[4] = ((in[2] >> 6) & 0x7f) | (in[3] << 7);
-        out[5] = in[3] >> 1;
-        out[6] = ((in[3] >> 9) & 0x0f) | (in[4] << 4);
-        out[7] = in[4] >> 4;
-        out[8] = ((in[4] >> 12) & 0x01) | (in[5] << 1);
-        out[9] = ((in[5] >> 7) & 0x3f) | (in[6] << 6);
-        out[10] = in[6] >> 2;
-        out[11] = ((in[6] >> 10) & 0x07) | (in[7] << 3);
-        out[12] = in[7] >> 5;
+        out[0] = (uint8_t) (in[0]);
+        out[1] = (uint8_t) (((in[0] >> 8) & 0x1f) | (in[1] << 5));
+        out[2] = (uint8_t) (in[1] >> 3);
+        out[3] = (uint8_t) (((in[1] >> 11) & 0x03) | (in[2] << 2));
+        out[4] = (uint8_t) (((in[2] >> 6) & 0x7f) | (in[3] << 7));
+        out[5] = (uint8_t) (in[3] >> 1);
+        out[6] = (uint8_t) (((in[3] >> 9) & 0x0f) | (in[4] << 4));
+        out[7] = (uint8_t) (in[4] >> 4);
+        out[8] = (uint8_t) (((in[4] >> 12) & 0x01) | (in[5] << 1));
+        out[9] = (uint8_t) (((in[5] >> 7) & 0x3f) | (in[6] << 6));
+        out[10] = (uint8_t) (in[6] >> 2);
+        out[11] = (uint8_t) (((in[6] >> 10) & 0x07) | (in[7] << 3));
+        out[12] = (uint8_t) (in[7] >> 5);
         in += 8;
         out += 13;
     }
@@ -74,11 +74,11 @@ static void POLp2BS(uint8_t bytes[SABER_POLYCOMPRESSEDBYTES], const poly *data) 
     const uint16_t *in = data->coeffs;
     uint8_t *out = bytes;
     for (j = 0; j < SABER_N / 4; j++) {
-        out[0] = in[0];
-        out[1] = ((in[0] >> 8) & 0x03) | (in[1] << 2);
-        out[2] = ((in[1] >> 6) & 0x0f) | (in[2] << 4);
-        out[3] = ((in[2] >> 4) & 0x3f) | (in[3] << 6);
-        out[4] = in[3] >> 2;
+        out[0] = (uint8_t) (in[0]);
+        out[1] = (uint8_t) (((in[0] >> 8) & 0x03) | (in[1] << 2));
+        out[2] = (uint8_t) (((in[1] >> 6) & 0x0f) | (in[2] << 4));
+        out[3] = (uint8_t) (((in[2] >> 4) & 0x3f) | (in[3] << 6));
+        out[4] = (uint8_t) (in[3] >> 2);
         in += 4;
         out += 5;
     }
